Switch from express-session to cookie-session
This should cause no issues, sessions are used through passport, which
stored minimal amounts of data in the session, and express-session needs
a real db in production (even if for our use-case, using memory would be
fine). Switching to cookie-session means the session is entirely
client-side, avoiding the need for another db.

Signed-off-by: James Tocknell <aragilar@gmail.com>