Clean up user model for multiple providers
Provider profiles will be stored in their own space, with required
information extracted from from the profile as put in the top-level
area.

New providers need to put a file in `src/auth_providers`, which has a
function which returns a passport strategy ready for use, including
parsing profile information to create/update a user in the database.

Signed-off-by: James Tocknell <aragilar@gmail.com>